backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
faiss-cpu>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0

